chore: refactor-oc-table-cell-head-to-setup-script
