+++ packages/design-system/src/components/OcTable/OcTable.vue
@@ -100,8 +100,8 @@
 import OcThead from '../_OcTableHeader/_OcTableHeader.vue'
 import OcTbody from '../OcTableBody/OcTableBody.vue'
 import OcTr from '../_OcTableRow/_OcTableRow.vue'
-import OcTh from '../_OcTableCellHead/_OcTableCellHead.vue'
 import OcTd from '../OcTableCellData/OcTableCellData.vue'
+import OcTh from '../OcTableCellHead/OcTableCellHead.vue'
 import OcButton from '../OcButton/OcButton.vue'
 import { getSizeClass, Item, FieldType } from '../../helpers'
 import { computed } from 'vue'
