+++ packages/design-system/src/components/index.ts
@@ -42,8 +42,8 @@ export { default as OcTable } from './OcTable/OcTable.vue'
 export { default as OcTableCell } from './OcTableCell/OcTableCell.vue'
 export { default as OcTableBody } from './OcTableBody/OcTableBody.vue'
 export { default as OcTableCellData } from './OcTableCellData/OcTableCellData.vue'
-export { default as OcTableCellHead } from './_OcTableCellHead/_OcTableCellHead.vue'
 export { default as OcTableFooter } from './OcTableFooter/OcTableFooter.vue'
+export { default as OcTableCellHead } from './OcTableCellHead/OcTableCellHead.vue'
 export { default as OcTableHeader } from './_OcTableHeader/_OcTableHeader.vue'
 export { default as OcTableRow } from './_OcTableRow/_OcTableRow.vue'
 export { default as OcTableSimple } from './OcTableSimple/OcTableSimple.vue'
